refactor: verse list 구현
