+++ lib/views/verse_list_view.dart
@@ -5,11 +5,12 @@
 // import 'package:bibletree/views/empty_view.dart';
 // import 'package:implicitly_animated_list/implicitly_animated_list.dart';
 // import 'package:bibletree/main.dart';
-import 'package:bibletree/models/verse_singleton.dart';
-import 'package:bibletree/models/record_item.dart';
+// import 'package:bibletree/models/verse_singleton.dart';
+// import 'package:bibletree/models/record_item.dart';
+import 'package:bibletree/models/verse.dart';
 import 'package:bibletree/viewModels/verse_list_view_model.dart';
 import 'package:bibletree/views/empty_view.dart';
-import 'package:bibletree/views/verse_list_card.dart';
+import 'package:bibletree/widgets/verse_list_card.dart';
 import 'package:flutter/material.dart';
 import 'package:implicitly_animated_list/implicitly_animated_list.dart';
 import 'package:provider/provider.dart';
@@ -24,6 +25,7 @@ class VerseListView extends StatefulWidget {
 class _VerseListViewState extends State<VerseListView> {
   // Data related variables
   // final RecordBloc _recordBloc = RecordBloc(RecordRepository(RecordDao()));
+  Verse verse = Verse(id: 0, verse: 'verse', book: 'book', chapter: 'chapter');
 
   @override
   Widget build(BuildContext context) {
@@ -47,11 +49,21 @@ class _VerseListViewState extends State<VerseListView> {
         ],
       ),
       body: SafeArea(
-          child: verseListViewModel.records.isEmpty
-              ? const EmptyView()
-              : ImplicitlyAnimatedList(
-                  itemData: verseListViewModel.records,
-                  itemBuilder: (_, record) => Container())),
+        child: verseListViewModel.records.isEmpty
+            ? const EmptyView()
+            : ImplicitlyAnimatedList(
+                itemData: verseListViewModel.records,
+                itemBuilder: (_, record) => Container(
+                  margin: const EdgeInsets.fromLTRB(24, 6, 24, 6),
+                  child: VerseListCard(
+                    verse: verse,
+                    onTap: () => verseListViewModel.onTapVerseListCard(),
+                  ),
+                ),
+                itemEquality: (a, b) => a['verseId'] == b['verseId'],
+              ),
+      ),
+
       // SafeArea(
       //   child: StreamBuilder(
       //     stream: _recordBloc.recordListStream,
@@ -80,21 +92,21 @@ class _VerseListViewState extends State<VerseListView> {
     );
   }
 
-  Widget listItem(RecordItem record) {
-    return Container(
-      margin: const EdgeInsets.fromLTRB(24, 6, 24, 6),
-      child: GestureDetector(
-        behavior: HitTestBehavior.translucent,
-        onTap: () {
-          // Navigator.of(context)
-          //     .push(MaterialPageRoute(
-          //         builder: (context) => RecordView(
-          //             record, VerseSingleton.instance.list[record.verseId])))
-          //     .then((value) => _recordBloc.getRecordList());
-        },
-        child:
-            VerseListTile(verse: VerseSingleton.instance.list[record.verseId]),
-      ),
-    );
-  }
+  // Widget listItem(RecordItem record) {
+  //   return Container(
+  //     margin: const EdgeInsets.fromLTRB(24, 6, 24, 6),
+  //     child: GestureDetector(
+  //       behavior: HitTestBehavior.translucent,
+  //       onTap: () {
+  //         // Navigator.of(context)
+  //         //     .push(MaterialPageRoute(
+  //         //         builder: (context) => RecordView(
+  //         //             record, VerseSingleton.instance.list[record.verseId])))
+  //         //     .then((value) => _recordBloc.getRecordList());
+  //       },
+  //       child:
+  //           VerseListTile(verse: VerseSingleton.instance.list[record.verseId]),
+  //     ),
+  //   );
+  // }
 }
